requirements.txt
asgiref==3.3.4
Django==3.2.4
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.11.2
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
